feat: 상품 api 수정- 메소드 분리및 필드 타입 변경

+++ src/main/java/com/order/rush_order/product/dto/ProductDetailResponseDto.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class ProductDetailResponseDto {
     private Long id;
     private String name;
-    private float price;
+    private BigDecimal price;
     private String description;
     private long stock;
